Get country details from API instead of locally
Now when a CountryCard component (link to the specific country details
page) is pressed, instead of locally using a json to search for the
country details, the data is fetched from the API.

Additionally, there is an error and loading state in the country details'
page, to indicate the user if the details are loading or an error
occurred when fetching the data.

Also some changes to the formatCountryDetailsData function, because when
using the endpoint https://restcountries.com/v3.1/alpha/{code}, the
result can have some missing properties in the object and can cause some
errors when processing the data.